+++ packages/openapi-parser/src/utils/upgrade.ts
@@ -1,10 +1,19 @@
 import type { OpenAPIV3_1 } from '@scalar/openapi-types'
 
+import type { UnknownObject } from '@scalar/types/utils'
 import type { AnyObject, Filesystem, UpgradeResult } from '../types/index'
 import { getEntrypoint } from './getEntrypoint'
-import { makeFilesystem } from './makeFilesystem'
+import { isFilesystem } from './isFilesystem'
+import { normalize } from './normalize'
 import { upgradeFromThreeToThreeOne } from './upgradeFromThreeToThreeOne'
 import { upgradeFromTwoToThree } from './upgradeFromTwoToThree'
+
+const upgraders = [
+  // Swagger 2.0 -> OpenAPI 3.0
+  upgradeFromTwoToThree,
+  // OpenAPI 3.0 -> OpenAPI 3.1
+  upgradeFromThreeToThreeOne,
+]
 
 /**
  * Upgrade specification to OpenAPI 3.1.0
@@ -16,12 +25,11 @@ export function upgrade(value: string | AnyObject | Filesystem): UpgradeResult<O
       version: '3.1',
     }
   }
-  const upgraders = [upgradeFromTwoToThree, upgradeFromThreeToThreeOne]
 
   // TODO: Run upgrade over the whole filesystem
   const result = upgraders.reduce(
-    (currentSpecification, upgrader) => upgrader(currentSpecification),
-    getEntrypoint(makeFilesystem(value)).specification,
+    (currentSpecification, upgrader) => upgrader(currentSpecification as UnknownObject),
+    isFilesystem(value) ? getEntrypoint(value).specification : normalize(value),
   ) as OpenAPIV3_1.Document
 
   return {