+++ packages/snippetz/src/utils/create-search-params.ts
@@ -1,0 +1,10 @@
+import type { HarRequest } from '@scalar/types/snippetz'
+
+/** Creates a new URL search params object and sets query params so we can handle arrays */
+export const createSearchParams = (query: HarRequest['queryString'] = []) => {
+  const searchParams = new URLSearchParams()
+  query.forEach((q) => {
+    searchParams.append(q.name, q.value)
+  })
+  return searchParams
+}